chore(tests): colocate frontend tests
