+++ CLAUDE.md
@@ -32,6 +32,7 @@ Rules:
 - Numbers **never get reused** — removed scenarios leave gaps. New scenarios take the next free number.
 - Every test (Rust, Vitest) MUST be traceable to one or more scenario IDs, surfaced where the runner prints them on failure. For Vitest, use the literal bracketed form in `it(...)`: `it('[IT-005] add intake creates entry', ...)`. For Rust, call the `scenario!` macro (defined in `src-tauri/src/test_support.rs`) as the first line of the test body — `scenario!("[IT-021]", "[VAL-007]");` — and let the function name describe the _behavior_, not the ID: `fn amount_below_lower_bound_rejected()`. nextest captures the macro's stdout per-test and prints it on failure, so every cited ID surfaces in CI output. See `_conv-test-traceability` for the authoritative rule.
 - Concrete values (ranges, lengths, formats) live **inside scenarios**, not inside the SHALL prose. A scenario is a test case; the SHALL describes the behavior. Keep SHALLs behavioral and short.
+- **Test layout.** Vitest tests live **colocated** as `<name>.test.ts` (or `<name>.svelte.test.ts` when the test uses runes) next to the source file they exercise — e.g. `src/lib/api/category.test.ts` next to `category.ts`, `src/lib/component/intake/IntakeMask.test.ts` next to `IntakeMask.svelte`. For SvelteKit route files, **drop the `+` prefix** on the test (`src/routes/welcome/page.test.ts`, not `+page.test.ts`); the `+` prefix is reserved by SvelteKit and breaks the dev server. Shared test infrastructure (`TestWrapper.svelte`, mocks, polyfills, `setup.ts`) stays under `tests/`. Rust integration tests follow the standard `src-tauri/tests/<cmd|repo>/test_*.rs` layout.
 
 The current prefix registry:
