--- tests/lib/component/Body.test.ts
@@ -1,73 +1,0 @@
-import { describe, it, expect } from 'vitest';
-import { render } from '@testing-library/svelte';
-import Body from '../../../src/lib/component/wizard/body/Body.svelte';
-import type { LibreUser, WizardInput } from '$lib/api/gen';
-import { CalculationGoalSchema, CalculationSexSchema } from '$lib/api/gen';
-import { setupVeilchenMock } from '../../utils/mocks';
-
-// Setup common mocks
-setupVeilchenMock();
-
-describe('Body Component', () => {
-  const mockWizardInput: WizardInput = {
-    age: 25,
-    height: 175,
-    weight: 70,
-    sex: CalculationSexSchema.enum.MALE,
-    activityLevel: 1.5,
-    weeklyDifference: 0.7,
-    calculationGoal: CalculationGoalSchema.enum.LOSS
-  };
-
-
-  const mockUserData: LibreUser = {
-    id: 1,
-    name: 'Arnie',
-    avatar: 'abc'
-  }
-
-  it('should render component without crashing', () => {
-    const { container } = render(Body, {
-      props: {
-        wizardInput: mockWizardInput,
-        userData: mockUserData
-      }
-    });
-
-    expect(container).toBeDefined();
-  });
-
-  it('should apply correct container styling', () => {
-    const { container } = render(Body, {
-      props: {
-        wizardInput: mockWizardInput,
-        userData: mockUserData
-      }
-    });
-
-    const wrapper = container.querySelector('.p-4');
-    expect(wrapper).toBeDefined();
-
-    const fieldset = container.querySelector('.fieldset');
-    expect(fieldset).toBeDefined();
-  });
-
-  it('should handle different wizard input values', () => {
-    const differentInput: WizardInput = {
-      ...mockWizardInput,
-      age: 30,
-      height: 180,
-      weight: 75,
-      sex: CalculationSexSchema.enum.FEMALE
-    };
-
-    const { container } = render(Body, {
-      props: {
-        wizardInput: differentInput,
-        userData: mockUserData
-      }
-    });
-
-    expect(container).toBeDefined();
-  });
-});
